Updated the code in different files:
- round_icon_button.dart
Made the RawMaterialButton to ElevatedButton as this was deprecated. Also made the class Stateful. This was done to add a long press action for the button using GestureDetector.

- results_page.dart
Added an animated sticker to be displayed above the bmi value according to the BMI of the user.

- calculator_brain.dart
defined a new function to return the corresponding animated sticker according to bmi result result.

- input_page.dart
Upon Long press of + and - buttons now updates the value continuously without the need to keep pressing the buttons again and again.
If the height card is tapped then the unit of height changes from cm to ft. 
Also on pressing the - button continuously decreases the value even below 0 which is not acceptable as there is nothing like negative weight.